// src/api-client/models/PlannedExpenseViewDto.ts
export interface PlannedExpenseViewDto {
    /**
     * 
     * @type {string}
     * @memberof PlannedExpenseViewDto
     */
    id?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PlannedExpenseViewDto
     */
    userId?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PlannedExpenseViewDto
     */
    itemType?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PlannedExpenseViewDto
     */
    name?: string | null;
    /**
     * 
     * @type {Date}
     * @memberof PlannedExpenseViewDto
     */
    date?: Date | null;
    /**
     * 
     * @type {number}
     * @memberof PlannedExpenseViewDto
     */
    dayOfMonth?: number | null;
    /**
     * 
     * @type {boolean}
     * @memberof PlannedExpenseViewDto
     */
    isRecurring?: boolean;
    /**
     * 
     * @type {number}
     * @memberof PlannedExpenseViewDto
     */
    totalAmount?: number;
    /**
     * 
     * @type {string}
     * @memberof PlannedExpenseViewDto
     */
    categoryId?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PlannedExpenseViewDto
     */
    categoryName?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PlannedExpenseViewDto
     */
    categoryImage?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PlannedExpenseViewDto
     */
    categoryColor?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PlannedExpenseViewDto
     */
    subCategory?: string | null;
    /**
     * 
     * @type {Date}
     * @memberof PlannedExpenseViewDto
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof PlannedExpenseViewDto
     */
    updatedAt?: Date;
    /**
     * 
     * @type {number}
     * @memberof PlannedExpenseViewDto
     */
    amountSpent?: number;
    /**
     * 
     * @type {number}
     * @memberof PlannedExpenseViewDto
     */
    percentageSpent?: number;
    /**
     * 
     * @type {number}
     * @memberof PlannedExpenseViewDto
     */
    amountLeft?: number;
    /**
     * 
     * @type {boolean}
     * @memberof PlannedExpenseViewDto
     */
    readonly isCompleted?: boolean;
}